backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
python-dotenv==1.0.0
gunicorn==20.1.0
requests==2.26.0
webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==9.0.0
requests==2.26.0
PyJWT==2.1.0
djangorestframework-simplejwt==4.7.2
django-filter==23.2